+++ src/utils.ts
@@ -6,28 +6,35 @@ import GObject from 'gi://GObject';
 export const USER = GLib.get_user_name();
 export const CACHE_DIR = `${GLib.get_user_cache_dir()}/${pkg.name.split('.').pop()}`;
 
-export function readFile(path: string) {
+export function readFile(file: string | Gio.File) {
     try {
-        const f = Gio.File.new_for_path(path);
+        const f = typeof file === 'string'
+            ? Gio.File.new_for_path(file)
+            : file;
+
         const [, bytes] = f.load_contents(null);
         return new TextDecoder().decode(bytes);
     } catch (_) {
-        // no file
         return '';
     }
 }
 
-export function readFileAsync(path: string): Promise<string> {
-    const file = Gio.File.new_for_path(path);
+export function readFileAsync(file: string | Gio.File): Promise<string> {
+    const f = typeof file === 'string'
+        ? Gio.File.new_for_path(file)
+        : file;
 
     return new Promise((resolve, reject) => {
-        file.load_contents_async(null, (_, res) => {
+        f.load_contents_async(null, (_, res) => {
             try {
-                const [success, bytes] = file.load_contents_finish(res);
-                return success
-                    ? resolve(new TextDecoder().decode(bytes))
-                    : reject(Error(
-                        `reading file ${path} was unsuccessful`));
+                const [success, bytes] = f.load_contents_finish(res);
+                if (success) {
+                    resolve(new TextDecoder().decode(bytes));
+                }
+                else {
+                    const path = typeof file === 'string' ? file : file.get_path();
+                    reject(Error(`reading file ${path} was unsuccessful`));
+                }
             } catch (error) {
                 reject(error);
             }
@@ -57,6 +64,35 @@ export function writeFile(string: string, path: string): Promise<Gio.File> {
     });
 }
 
+const fileMonitors: Map<Gio.FileMonitor, boolean> = new Map;
+export function monitorFile(
+    path: string,
+    callback?: (file: Gio.File, event: Gio.FileMonitorEvent) => void,
+    type: 'file' | 'directory' = 'file',
+    flags = Gio.FileMonitorFlags.NONE,
+) {
+    try {
+        const file = Gio.File.new_for_path(path);
+        const mon = type === 'directory'
+            ? file.monitor_directory(flags, null)
+            : file.monitor_file(flags, null);
+
+        if (callback)
+            mon.connect('changed', (_, file, _f, event) => callback(file, event));
+
+        // we need to save a reference in case the user doesn't
+        // otherwise GC will pick it up
+        fileMonitors.set(mon, true);
+        mon.connect('notify::cancelled', () => {
+            fileMonitors.delete(mon);
+        });
+
+        return mon;
+    } catch (error) {
+        logError(error);
+        return null;
+    }
+}
 
 export function loadInterfaceXML(iface: string) {
     const uri = `resource:///com/github/Aylur/ags/dbus/${iface}.xml`;